Rename common.py->dataobject.py
