+++ pyvista/core/composite.py
@@ -14,7 +14,7 @@ from vtk import vtkMultiBlockDataSet
 
 import pyvista
 from pyvista.utilities import get_array, is_pyvista_dataset, wrap
-from .common import DataObject, Common
+from .dataobject import DataObject, Common
 from .filters import CompositeFilters
 from .._typing import Vector
