+++ pyvista/core/objects.py
@@ -9,7 +9,7 @@ import vtk
 import pyvista
 from pyvista.utilities import (FieldAssociation, assert_empty_kwargs, get_array,
                                row_array)
-from .common import DataObject
+from .dataobject import DataObject
 from .datasetattributes import DataSetAttributes
